requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.2
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-nested-inline==0.4.4
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
simplejson==3.17.6
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
